Added the Rate App Extra Functionalities
I added the functionalities of my demo that was created in the older version and combined it with the style of the newer version. Added a new look to the page to look more like Figma. White text box, App Store button and changed the UI structure for better user experience.